+++ x/bank/keeper/genesis_test.go
@@ -44,7 +44,6 @@ func (suite *IntegrationTestSuite) getTestBalances() []types.Balance {
 }
 
 func (suite *IntegrationTestSuite) TestInitGenesis() {
-	require := suite.Require()
 	m := types.Metadata{Description: sdk.DefaultBondDenom, Base: sdk.DefaultBondDenom, Display: sdk.DefaultBondDenom}
 	g := types.DefaultGenesisState()
 	g.DenomMetadata = []types.Metadata{m}
@@ -52,5 +51,5 @@ func (suite *IntegrationTestSuite) TestInitGenesis() {
 	bk.InitGenesis(suite.ctx, g)
 
 	m2 := bk.GetDenomMetaData(suite.ctx, m.Base)
-	require.Equal(m, m2)
+	suite.Require().Equal(m, m2)
 }